+++ eng/versioning/version_client.txt
@@ -176,7 +176,7 @@ com.azure.spring:azure-spring-integration-test;2.6.0;2.7.0-beta.1
 com.azure.spring:azure-spring-boot-test-aad;1.0.0;1.0.0
 com.azure.spring:azure-spring-boot-test-aad-b2c;1.0.0;1.0.0
 com.azure.spring:azure-spring-boot-test-selenium-common;1.0.0;1.0.0
-com.azure.spring:azure-spring-boot-test-aad-obo;1.0.0;1.0.0
+com.azure.spring:azure-spring-boot-test-aad-webapp-and-resource-server-with-obo;1.0.0;1.0.0
 com.azure.spring:azure-spring-boot-test-aad-resource-server;1.0.0;1.0.0
 com.azure.spring:azure-spring-boot-test-aad-resource-server-by-filter;1.0.0;1.0.0
 com.azure.spring:azure-spring-boot-test-core;1.0.0;1.0.0
