+++ sdk/spring/azure-spring-boot-test-aad-webapp-and-resource-server-with-obo/src/test/java/com/azure/test/aad/selenium/AADWebAppAndWebApiInOneAppIT.java
@@ -5,6 +5,8 @@ package com.azure.test.aad.selenium;
 import com.azure.spring.aad.webapi.AADResourceServerWebSecurityConfigurerAdapter;
 import com.azure.spring.aad.webapp.AADWebSecurityConfigurerAdapter;
 import com.azure.spring.test.aad.AADWebApiITHelper;
+import com.azure.spring.utils.AzureCloudUrls;
+import com.azure.test.aad.common.AADSeleniumITHelper;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
@@ -33,14 +35,16 @@ import org.springframework.web.reactive.function.client.WebClient;
 
 import java.security.Principal;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 
 import static com.azure.spring.test.Constant.MULTI_TENANT_SCOPE_GRAPH_READ;
 import static com.azure.spring.test.EnvironmentVariable.AAD_MULTI_TENANT_CLIENT_ID;
 import static com.azure.spring.test.EnvironmentVariable.AAD_MULTI_TENANT_CLIENT_SECRET;
+import static com.azure.spring.test.EnvironmentVariable.AAD_TENANT_ID_1;
 import static com.azure.spring.test.EnvironmentVariable.AAD_USER_NAME_2;
 import static com.azure.spring.test.EnvironmentVariable.AAD_USER_PASSWORD_2;
-import static com.azure.test.aad.selenium.AADSeleniumITHelper.createDefaultProperties;
+import static com.azure.spring.test.EnvironmentVariable.AZURE_CLOUD_TYPE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction.oauth2AuthorizedClient;
@@ -55,7 +59,20 @@ public class AADWebAppAndWebApiInOneAppIT {
 
     @BeforeAll
     public void beforeAll() {
-        properties = createDefaultProperties();
+        properties = new HashMap<>();
+        properties.put("azure.activedirectory.tenant-id", AAD_TENANT_ID_1);
+        properties.put("azure.activedirectory.client-id", AAD_MULTI_TENANT_CLIENT_ID);
+        properties.put("azure.activedirectory.client-secret", AAD_MULTI_TENANT_CLIENT_SECRET);
+        properties.put("azure.activedirectory.user-group.allowed-groups", "group1");
+        properties.put("azure.activedirectory.post-logout-redirect-uri", "http://localhost:${server.port}");
+        properties.put("azure.activedirectory.base-uri", AzureCloudUrls.getBaseUrl(AZURE_CLOUD_TYPE));
+        properties.put("azure.activedirectory.graph-base-uri", AzureCloudUrls.getGraphBaseUrl(AZURE_CLOUD_TYPE));
+        properties.put("azure.activedirectory.application-type", "web_application_and_resource_server");
+
+        properties.put("azure.activedirectory.app-id-uri", "api://" + AAD_MULTI_TENANT_CLIENT_ID);
+        properties.put("azure.activedirectory.authorization-clients.graph.scopes",
+            "https://graph.microsoft.com/User.Read");
+        properties.put("azure.activedirectory.authorization-clients.graph.authorization-grant-type", "on_behalf_of");
     }
 
     @Test
