+++ sdk/spring/azure-spring-boot-test-aad-webapp-and-resource-server-with-obo/src/test/java/com/azure/test/aad/webapi/obo/AADWebApiOboIT.java
