+++ sdk/spring/azure-spring-boot-test-aad/src/test/java/com/azure/test/aad/selenium/conditional/access/AADConditionalAccessIT.java
@@ -3,7 +3,7 @@
 
 package com.azure.test.aad.selenium.conditional.access;
 
-import com.azure.test.aad.selenium.AADSeleniumITHelper;
+import com.azure.test.aad.common.AADSeleniumITHelper;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -25,13 +25,13 @@ import org.springframework.web.reactive.function.client.WebClient;
 import java.util.Map;
 
 import static com.azure.spring.test.EnvironmentVariable.CONDITIONAL_ACCESS_POLICY_TEST_WEB_API_A_CLIENT_ID;
-import static com.azure.test.aad.selenium.AADSeleniumITHelper.createDefaultProperties;
+import static com.azure.test.aad.selenium.AADITHelper.createDefaultProperties;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction.oauth2AuthorizedClient;
 
 
 /**
- * Before running this test case, follow the <a href="https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/spring/azure-spring-boot-test-aad/README.md#what-is-aad-conditional-access-in-aad">guide</a> to start <a href="https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/spring/azure-spring-boot-test-aad-obo/src/test/java/com/azure/test/aad/webapi/conditional/access/webapi/a/ConditionalAccessPolicyTestWebApiA.java">WebapiA</a> and  <a href="https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/spring/azure-spring-boot-test-aad-resource-server/src/test/java/com/azure/test/aad/conditional/access/webapi/b/ConditionalAccessPolicyTestWebApiB.java">webapiB</a>.
+ * Before running this test case, follow the <a href="https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/spring/azure-spring-boot-test-aad/README.md#what-is-aad-conditional-access-in-aad">guide</a> to start <a href="https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/spring/azure-spring-boot-test-aad-webapp-and-resource-server-with-obo/src/test/java/com/azure/test/aad/webapi/conditional/access/webapi/a/ConditionalAccessPolicyTestWebApiA.java">WebapiA</a> and  <a href="https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/spring/azure-spring-boot-test-aad-resource-server/src/test/java/com/azure/test/aad/conditional/access/webapi/b/ConditionalAccessPolicyTestWebApiB.java">webapiB</a>.
  */
 @Disabled
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
