+++ sdk/spring/pom.xml
@@ -141,7 +141,7 @@
         <module>azure-spring-boot-starter-storage</module>
         <module>azure-spring-boot-test-aad</module>
         <module>azure-spring-boot-test-aad-b2c</module>
-        <module>azure-spring-boot-test-aad-obo</module>
+        <module>azure-spring-boot-test-aad-webapp-and-resource-server-with-obo</module>
         <module>azure-spring-boot-test-aad-resource-server</module>
         <module>azure-spring-boot-test-aad-resource-server-by-filter</module>
         <module>azure-spring-boot-test-application</module>
