+++ sdk/spring/spring-test-template.yml
@@ -26,7 +26,7 @@ parameters:
     - name: azure-spring-boot-test-aad-b2c
       groupId: com.azure.spring
       safeName: azurespringboottestaadb2c
-    - name: azure-spring-boot-test-aad-obo
+    - name: azure-spring-boot-test-aad-webapp-and-resource-server-with-obo
       groupId: com.azure.spring
       safeName: azurespringboottestaadobo
     - name: azure-spring-boot-test-aad-resource-server
